requirements/default.txt
# BharatVerse - Complete Requirements
# Cultural Heritage Preservation Platform

#=============================================================================
# CORE DEPENDENCIES (Required)
#=============================================================================
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0

#=============================================================================
# WEB FRAMEWORK & API
#=============================================================================
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6

#=============================================================================
# DATABASE & STORAGE
#=============================================================================
# PostgreSQL database adapter
psycopg2-binary

# Redis caching
redis

# MinIO object storage (S3-compatible)
minio

# SQL toolkit and ORM
sqlalchemy>=2.0.23

# Environment configuration
python-dotenv>=1.0.0

#=============================================================================
# DATA VISUALIZATION & UI
#=============================================================================
plotly>=5.17.0
altair>=5.0.0
streamlit-option-menu>=0.3.6
streamlit-lottie>=0.0.5
streamlit-aggrid>=0.3.4

# Charts and graphs
seaborn>=0.12.0
matplotlib>=3.7.0
# wordcloud>=1.9.0  # Commented out due to build issues in Docker

#=============================================================================
# UTILITY LIBRARIES
#=============================================================================
requests>=2.31.0
beautifulsoup4>=4.12.0
markdown>=3.5.0

# Data validation and security
pydantic>=2.5.0
cryptography>=41.0.0

# OAuth and Authentication
authlib>=1.2.1
httpx>=0.25.0
python-jose[cryptography]>=3.3.0

# Performance and caching
diskcache>=5.6.0

#=============================================================================
# AUDIO PROCESSING (Now Required for Real AI Features)
#=============================================================================
sounddevice>=0.4.6
soundfile>=0.12.1
librosa>=0.10.0
pydub>=0.25.1

#=============================================================================
# MACHINE LEARNING & AI (Now Required for Real AI Features)
#=============================================================================
transformers>=4.36.0
torch>=2.1.0
torchvision>=0.16.0
sentence-transformers>=2.2.2
openai-whisper>=20231117

#=============================================================================
# TEXT PROCESSING & NLP (Now Required for Real AI Features)
#=============================================================================
langdetect>=1.0.9
textstat>=0.7.3

#=============================================================================
# INDIAN LANGUAGE SUPPORT (Enhanced)
#=============================================================================
# indic-nlp-library>=0.92  # Optional for advanced Indic language support
# polyglot>=16.7.4  # Optional for multilingual support

#=============================================================================
# IMAGE PROCESSING (Now Required for Real AI Features)
#=============================================================================
opencv-python>=4.8.0
scikit-image>=0.21.0

#=============================================================================
# INSTALLATION NOTES
#=============================================================================
# 1. Basic installation (core features only):
#    pip install streamlit pandas numpy Pillow plotly altair psycopg2-binary redis minio
#
# 2. Full installation (all features):
#    pip install -r requirements.txt
#
# 3. Development installation (with optional packages):
#    Uncomment the optional sections above before installing
#
# 4. Production installation:
#    Use the basic installation for minimal dependencies
#
# 5. Docker installation:
#    All dependencies are included in the Docker image
langdetect
textblob
wordcloud